:recycle: DRY StUF-ZDS test code
Moved test utilities and base test case with assertions to the stuf
package.
