+++ src/openforms/registrations/contrib/stuf_zds/tests/test_backend.py
@@ -2,13 +2,8 @@ import dataclasses
 from decimal import Decimal
 from unittest.mock import patch
 
-from django.template import loader
-from django.test import TestCase
-
 import requests_mock
 from freezegun import freeze_time
-from lxml import etree
-from lxml.etree import ElementTree
 from privates.test import temp_private_root
 
 from openforms.logging.models import TimelineLogProxy
@@ -17,8 +12,9 @@ from openforms.submissions.tests.factories import (
     SubmissionFileAttachmentFactory,
 )
 from openforms.submissions.tests.mixins import VariablesTestMixin
-from stuf.stuf_zds.client import nsmap
 from stuf.stuf_zds.models import StufZDSConfig
+from stuf.stuf_zds.tests import StUFZDSTestBase
+from stuf.stuf_zds.tests.utils import load_mock, match_text, xml_from_request_history
 from stuf.tests.factories import StufServiceFactory
 
 from ....constants import RegistrationAttribute
@@ -26,82 +22,7 @@ from ....service import extract_submission_reference
 from ..plugin import PartialDate, StufZDSRegistration
 
 
-def load_mock(name, context=None):
-    return loader.render_to_string(
-        f"stuf_zds/soap/response-mock/{name}", context
-    ).encode("utf8")
-
-
-def match_text(text):
-    # requests_mock matcher for SOAP requests
-    def _matcher(request):
-        return text in (request.text or "")
-
-    return _matcher
-
-
-def xml_from_request_history(m, index) -> ElementTree:
-    request = m.request_history[index]
-    xml = etree.fromstring(bytes(request.text, encoding="utf8"))
-    return xml
-
-
-class StufTestBase(TestCase):
-    namespaces = nsmap
-
-    def assertXPathExists(self, xml_doc, xpath):
-        elements = xml_doc.xpath(xpath, namespaces=self.namespaces)
-        if len(elements) == 0:
-            self.fail(f"cannot find XML element(s) with xpath {xpath}")
-
-    def assertXPathNotExists(self, xml_doc, xpath):
-        elements = xml_doc.xpath(xpath, namespaces=self.namespaces)
-        if len(elements) != 0:
-            self.fail(
-                f"found {len(elements)} unexpected XML element(s) with xpath {xpath}"
-            )
-
-    def assertXPathCount(self, xml_doc, xpath, count):
-        elements = xml_doc.xpath(xpath, namespaces=self.namespaces)
-        self.assertEqual(
-            len(elements),
-            count,
-            f"cannot find exactly {count} XML element(s) with xpath {xpath}",
-        )
-
-    def assertXPathEquals(self, xml_doc, xpath, text):
-        elements = xml_doc.xpath(xpath, namespaces=self.namespaces)
-        self.assertGreaterEqual(
-            len(elements), 1, f"cannot find XML element(s) with xpath {xpath}"
-        )
-        self.assertEqual(
-            len(elements), 1, f"multiple XML element(s) found for xpath {xpath}"
-        )
-        if isinstance(elements[0], str):
-            self.assertEqual(elements[0].strip(), text, f"at xpath {xpath}")
-        else:
-            elem_text = elements[0].text
-            if elem_text is None:
-                elem_text = ""
-            else:
-                elem_text = elem_text.strip()
-            self.assertEqual(elem_text, text, f"at xpath {xpath}")
-
-    def assertXPathEqualDict(self, xml_doc, path_value_dict):
-        for path, value in path_value_dict.items():
-            self.assertXPathEquals(xml_doc, path, value)
-
-    def assertSoapXMLCommon(self, xml_doc):
-        self.assertIsNotNone(xml_doc)
-        self.assertXPathExists(
-            xml_doc, "/*[local-name()='Envelope']/*[local-name()='Header']"
-        )
-        self.assertXPathExists(
-            xml_doc, "/*[local-name()='Envelope']/*[local-name()='Body']"
-        )
-
-
-class StufZDSHelperTests(StufTestBase):
+class StufZDSHelperTests(StUFZDSTestBase):
     def test_partial_date(self):
         # good
         actual = PartialDate.parse("2020-01-01")
@@ -203,7 +124,7 @@ class StufZDSHelperTests(StufTestBase):
 @freeze_time("2020-12-22")
 @temp_private_root()
 @requests_mock.Mocker()
-class StufZDSPluginTests(VariablesTestMixin, StufTestBase):
+class StufZDSPluginTests(VariablesTestMixin, StUFZDSTestBase):
     """
     test the plugin function
     """
