+++ src/stuf/stuf_zds/tests/__init__.py
@@ -1,0 +1,1 @@
+from .testcases import StUFZDSTestBase  # noqa
