+++ src/stuf/stuf_zds/tests/testcases.py
@@ -1,0 +1,7 @@
+from stuf.tests.testcases import StUFTestBase
+
+from ..client import nsmap
+
+
+class StUFZDSTestBase(StUFTestBase):
+    namespaces = nsmap
