+++ src/stuf/stuf_zds/tests/utils.py
@@ -1,0 +1,24 @@
+from django.template import loader
+
+from lxml import etree
+from lxml.etree import ElementTree
+
+
+def load_mock(name, context=None):
+    return loader.render_to_string(
+        f"stuf_zds/soap/response-mock/{name}", context
+    ).encode("utf8")
+
+
+def match_text(text):
+    # requests_mock matcher for SOAP requests
+    def _matcher(request):
+        return text in (request.text or "")
+
+    return _matcher
+
+
+def xml_from_request_history(m, index) -> ElementTree:
+    request = m.request_history[index]
+    xml = etree.fromstring(bytes(request.text, encoding="utf8"))
+    return xml
